+++ lib/classes/rmq-error-handler.class.ts
@@ -1,28 +1,16 @@
-import { ERROR_TYPE } from '../constants';
-import { RMQTransportError } from './rmq-transport-error.class';
 import { RMQError } from './rmq-error.class';
+import { IRmqErrorHeaders } from '../interfaces/rmq-error-headers.interface';
+import { MessagePropertyHeaders } from 'amqplib';
 
 export class RMQErrorHandler {
-	public static handle(headers: any): any {
-		switch (headers['-x-type']) {
-			case ERROR_TYPE.TRANSPORT:
-				return new RMQTransportError(
-					headers['-x-error'],
-					headers['-x-status-code'],
-					headers['-x-data'],
-					headers['-x-service'],
-					headers['-x-host']
-				);
-				break;
-			case ERROR_TYPE.RMQ:
-				return new RMQError(
-					headers['-x-error'],
-					headers['-x-status-code'],
-					headers['-x-data'],
-					headers['-x-service'],
-					headers['-x-host']
-				);
-				break;
-		}
+	public static handle(headers: IRmqErrorHeaders | MessagePropertyHeaders): Error | RMQError {
+		return new RMQError(
+			headers['-x-error'],
+			headers['-x-type'],
+			headers['-x-status-code'],
+			headers['-x-data'],
+			headers['-x-service'],
+			headers['-x-host']
+		);
 	}
 }