+++ lib/classes/rmq-error.class.ts
@@ -1,25 +1,43 @@
+import { ERROR_TYPE } from '../constants';
+
 export class RMQError extends Error {
-	
+	/**
+	 * @summary Error message
+	 */
 	message: string;
 
-	/* Your custom error code */
-	code?: string;
+	/**
+	 * @summary Error code
+	 */
+	code?: number;
 
-	/* Error custom data */
+	/**
+	 * @summary Error custom data
+	 */
 	data?: any;
-	
-	/* Service custon name */
+
+	/**
+	 * @summary Service name
+	 */
 	service?: string;
-	
-	/* Host name */
+
+	/**
+	 * @summary Host name
+	 */
 	host?: string;
 
-	constructor(message: string, code?: string, data?: any, service?: string, host?: string) {
+	/**
+	 * @summary Host name
+	 */
+	type?: ERROR_TYPE;
+
+	constructor(message: string, type: ERROR_TYPE, code?: number, data?: any, service?: string, host?: string) {
 		super();
 		Object.setPrototypeOf(this, new.target.prototype);
+		this.message = message;
+		this.type = type;
 		this.code = code;
 		this.data = data;
-		this.message = message;
 		this.service = service;
 		this.host = host;
 	}