+++ lib/constants.ts
@@ -15,6 +15,7 @@ export const ERROR_TIMEOUT: string = 'Response timeout error';
 
 export const DEFAULT_RECONNECT_TIME: number = 5;
 export const DEFAULT_TIMEOUT: number = 30000;
+export const DEFAULT_PREFETCH_COUNT: number = 0;
 
 export enum ERROR_TYPE {
 	TRANSPORT = 'TRANSPORT',